+++ src/lib/types/storage/StorageType.ts
@@ -1,0 +1,9 @@
+enum StorageType {
+    SERVICE = 'service', //FFU
+    ACCESSORY = 'accessory', //FFU
+    HOST = 'host', //FFU
+    CUSTOM_CHARACTERISTICS = 'customCharacteristics',
+    OTHER = 'other', //FFU
+}
+
+export { StorageType }